Add controller for DataActivity and InstagramPost
Add Constant class with all constants data
