+++ app/src/main/java/com/example/instagramrecovery/presentation/Constants.java
@@ -1,0 +1,9 @@
+package com.example.instagramrecovery.presentation;
+
+public class Constants {
+    public static String BASE_URL1 = "https://api.instagram.com";
+    public static String BASE_URL2 = "https://graph.instagram.com";
+    public static String app_id = "598355621029954";
+    public static String app_secret = "9e1642c8c88dcdeada90d8505e3903e2";
+    public static String redirect_uri = "https://math-dzieu.github.io/Info.github.io";
+}
